Replace OFA package dependency with code stored locally (#15)
This change removes OFA as an external dependency. Instead, parts of the
OFA code that are needed to run LINAS search with OFA MBv3/ResNet50
supernets were copied from the original repository into the DyNAS-T.

The reason behind this is that OFA brings a set of dependency versions
that may cause issues when integrated with other libraries. Second
reason is that this move enables us to have more control over the
evaluation process, as now we can directly modify OFA's pipeline.

---------

Signed-off-by: Maciej Szankin <maciej.szankin@intel.com>