Bug: fix order of transformer-based supernet metrics
Following up on the PR#34, the issue remained that the ordering of
data in the CSV result file was not consistent across the DyNAS-T
search.

In this particular case, while the data was saved correctly to the CSV
file, it was not read properly by the predictor which caused various
issues in the search results.

Signed-off-by: Maciej Szankin <maciej.szankin@intel.com>