Fix OFA interface for ResNet50
The `get_subnet` method in `OFARunner` was using improper keys for
the resnet50 supernetwork.

To start with, it used `ks` which is only applicable to MobileNetV3
supernets and passed depth as it's value. Unused (`ks`) param will be
ignored, but the depth, as it was not set, will take the default value
of None, resulting in a limited search space.

Secondly, the depth param was set to width value. Valid values for both
depth and width are the same, so there was no raised error.

Signed-off-by: Maciej Szankin <maciej.szankin@intel.com>